requirements.txt
python-telegram-bot==20.7
pydub==0.25.1

